add Header to auth pages
